# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionIn": {
        "additionalProperties": false,
        "properties": {
          "dt": {
            "default": 1.0,
            "description": "Macro step, seconds.",
            "exclusiveMinimum": 0.0,
            "maximum": 60.0,
            "title": "Dt",
            "type": "number"
          },
          "revision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Default: the head.",
            "title": "Revision"
          },
          "scope": {
            "description": "Asset ids to simulate.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Scope",
            "type": "array"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "scope"
        ],
        "title": "SessionIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Graphene World Sim API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runtime/sessions": {
      "get": {
        "operationId": "list_sessions_api_runtime_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Sessions Api Runtime Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "runtime"
        ]
      },
      "post": {
        "operationId": "create_session_api_runtime_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/sessions/{sid}": {
      "delete": {
        "operationId": "delete_session_api_runtime_sessions__sid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "runtime"
        ]
      },
      "get": {
        "operationId": "get_session_api_runtime_sessions__sid__get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/sessions/{sid}/alarms": {
      "get": {
        "description": "Alarms raised by the scope's fault-alarm points, newest first, each with the trajectory\nevent that preceded it.",
        "operationId": "alarms_api_runtime_sessions__sid__alarms_get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "active"
              ],
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Alarms Api Runtime Sessions  Sid  Alarms Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alarms",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/sessions/{sid}/commands": {
      "post": {
        "operationId": "command_api_runtime_sessions__sid__commands_post",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/sessions/{sid}/conditions": {
      "put": {
        "description": "Change operating conditions: `dry_bulb_c`, `wet_bulb_c`, `relative_humidity`,\n`utility_available`, `it_fraction` ({room: fraction}).",
        "operationId": "conditions_api_runtime_sessions__sid__conditions_put",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Changes",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Conditions Api Runtime Sessions  Sid  Conditions Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Conditions",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/sessions/{sid}/diagnostics": {
      "get": {
        "description": "How the session is doing: partitions and their solvers, compile and step times, the\nreal-time factor, what the scope leaves out, and the last error.",
        "operationId": "diagnostics_api_runtime_sessions__sid__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Diagnostics Api Runtime Sessions  Sid  Diagnostics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnostics",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/sessions/{sid}/events": {
      "get": {
        "operationId": "events_api_runtime_sessions__sid__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Events Api Runtime Sessions  Sid  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "runtime"
        ]
      }
    },
...
    "/api/runtime/sessions/{sid}/frame": {
      "get": {
        "operationId": "frame_api_runtime_sessions__sid__frame_get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Frame Api Runtime Sessions  Sid  Frame Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frame",
        "tags": [
          "runtime"
        ]
      }
    },
...
  }
}